Zero out attachments instead of complete removal
Changes default "remove attachments" behavior to replace attachments
with an empty file with "Deleted:" prepended to the original filename.
This provides a clearer visual indication of removed attachments in the
context of the email and offers a lightweight and less conspicuous
alternative to using the "Add metadata" option. In addition, such emails
will still appear in results for a `has:attachment` search.

Also: Adds `.java-version` file, useful for those using jenv with
multiple versions of Java.

Closes rokstrnisa/unattach#39.